src: Several pages added
A form page, meter data page and a line data page was created.
The form page is a general input page for name and surname.
The meter data page displays meter data in a line graph.
The line data page shows randomly generated line data in a graph.
Navigation between the pages was added.

Signed-off-by: Arnold van der Merwe <arnold1.vandermerwe@gmail.com>